# openapi.yaml
openapi: 3.0.3
info:
  title: ERP/POS API
  description: |
    Sistema ERP de punto de venta y gestión de almacén.
    
    ## Características principales
    - Autenticación JWT con roles
    - Gestión de productos con variantes
    - Control de inventario en tiempo real
    - Punto de venta con precios automáticos (mayoreo/menudeo)
    - Facturación SAT preparada
    - Multi-caja y multi-usuario
    
    ## Arquitectura
    - Backend: FastAPI + PostgreSQL
    - Frontend: React
    - Base de datos con lógica de negocio centralizada
    
  version: 1.0.0
  contact:
    name: Soporte Técnico
    email: soporte@erp.com
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
  - url: https://api.erp.com/api/v1
    description: Producción
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido del endpoint /auth/login.
        Incluye en el header: `Authorization: Bearer {token}`

  # ============================================
  # ESQUEMAS
  # ============================================
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "STOCK_INSUFFICIENT"
        message:
          type: string
          example: "Stock insuficiente para completar la operación"
        details:
          type: object
          additionalProperties: true

    # --- Auth ---
...
    VentaCreateRequest:
      type: object
      required:
        - punto_venta_id
        - metodo_pago
      properties:
        punto_venta_id:
          type: integer
          example: 1
        cliente_nombre:
          type: string
          nullable: true
          example: "Juan Pérez"
        metodo_pago:
          type: string
          enum: [EFECTIVO, TARJETA, TRANSFERENCIA]
          example: "EFECTIVO"
    VentaCreateResponse:
      type: object
      properties:
        venta_id:
          type: integer
          example: 101
        folio:
          type: string
          example: "PV1-000101"
        fecha_creacion:
          type: string
          format: date-time
        estado:
          type: string
          enum: [ABIERTA, CERRADA, CANCELADA]
          example: "ABIERTA"
    VentaItemAddRequest:
      type: object
      required:
        - variante_id
        - cantidad
      properties:
        variante_id:
          type: integer
          example: 3
        cantidad:
          type: integer
          minimum: 1
          example: 5
        descuento_manual:
          type: number
          format: decimal
          minimum: 0
          example: 0.00
          description: Descuento adicional aplicado manualmente (requiere permisos)
    VentaItemAddResponse:
      type: object
      properties:
        detalle_id:
          type: integer
          example: 1
        variante_id:
          type: integer
          example: 3
        cantidad:
          type: integer
          example: 5
        precio_unitario_aplicado:
          type: number
          format: decimal
          example: 600.00
        regla_precio:
          type: string
          enum: [MENUDEO, MAYOREO]
          example: "MENUDEO"
        subtotal:
          type: number
          format: decimal
          example: 3000.00
        stock_restante:
          type: integer
          example: 37
        cantidad_total_producto_en_venta:
          type: integer
          example: 5
          description: Cantidad acumulada de este producto en toda la venta
    VentaDetalle:
      type: object
      properties:
        venta_id:
          type: integer
        folio:
          type: string
        fecha_venta:
          type: string
          format: date-time
        cliente_nombre:
          type: string
          nullable: true
        metodo_pago:
          type: string
        estado:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              detalle_id:
                type: integer
              nombre_producto:
                type: string
              nombre_variante:
                type: string
              cantidad:
                type: integer
              precio_unitario:
                type: number
              subtotal:
                type: number
        subtotal:
          type: number
        descuento_total:
          type: number
        total:
          type: number
        usuario:
          type: string
        punto_venta:
          type: string
    VentaCerrarResponse:
      type: object
      properties:
        venta_id:
          type: integer
        estado:
          type: string
          example: "CERRADA"
        total:
          type: number
          example: 3000.00
        ticket_generado:
          type: boolean
          example: true
        facturable:
          type: boolean
          example: true

    # --- Inventario ---
...
paths:
...
  /ventas:
    post:
      tags:
        - Ventas
      summary: Crear una nueva venta
      description: |
        Crea una venta nueva en estado ABIERTA.
        Esta es la primera operación del flujo POS.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VentaCreateRequest'
      responses:
        '201':
          description: Venta creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VentaCreateResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Ventas
      summary: Listar ventas
      description: Obtiene un listado de ventas con filtros opcionales
      security:
        - BearerAuth: []
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [ABIERTA, CERRADA, CANCELADA]
        - name: punto_venta_id
          in: query
          schema:
            type: integer
        - name: fecha_desde
          in: query
          schema:
            type: string
            format: date
        - name: fecha_hasta
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista de ventas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VentaDetalle'
  /ventas/{venta_id}:
    get:
      tags:
        - Ventas
      summary: Obtener detalle de una venta
      security:
        - BearerAuth: []
      parameters:
        - name: venta_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalle de la venta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VentaDetalle'
        '404':
          description: Venta no encontrada
  /ventas/{venta_id}/items:
    post:
      tags:
        - Ventas
      summary: Agregar producto a la venta
      description: |
        Agrega un producto (variante) a la venta.
        
        **La base de datos:**
        - Valida stock disponible
        - Aplica precio automático (mayoreo si cantidad total >= umbral)
        - Bloquea inventario
        - Calcula subtotal
        
        **Importante:** El precio NO viene del frontend, se calcula en PostgreSQL.
      security:
        - BearerAuth: []
      parameters:
        - name: venta_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VentaItemAddRequest'
      responses:
        '201':
          description: Producto agregado a la venta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VentaItemAddResponse'
        '409':
          description: Stock insuficiente o venta cerrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                stock_insuficiente:
                  value:
                    code: "STOCK_INSUFFICIENT"
                    message: "Stock insuficiente. Disponible: 2, solicitado: 5"
                venta_cerrada:
                  value:
                    code: "SALE_CLOSED"
                    message: "No se pueden agregar productos a una venta cerrada"
...
  /ventas/{venta_id}/cerrar:
    post:
      tags:
        - Ventas
      summary: Cerrar venta
      description: |
        Cierra la venta y ejecuta el proceso de finalización:
        - Descuenta inventario definitivamente
        - Genera ticket
        - Marca como facturable
        - Cambia estado a CERRADA
        
        **Importante:** Una vez cerrada, no se puede modificar.
      security:
        - BearerAuth: []
      parameters:
        - name: venta_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Venta cerrada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VentaCerrarResponse'
        '409':
          description: Venta ya cerrada o sin items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ventas/{venta_id}/cancelar:
    post:
      tags:
        - Ventas
      summary: Cancelar venta
      description: |
        Cancela una venta ABIERTA.
        Libera el inventario bloqueado.
      security:
        - BearerAuth: []
      parameters:
        - name: venta_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Venta cancelada
        '409':
          description: Venta ya cerrada (no se puede cancelar)

  # ============================================
  # PRODUCTOS
  # ============================================
...